move ParseRequest function to util package
